+++ Sources/HealthCardControl/Operations/CardChannelType+Version.swift
@@ -15,23 +15,9 @@
 //
 
 import CardReaderProviderApi
+import Combine
 import Foundation
 import HealthCardAccess
-
-enum CardAid: ApplicationIdentifier {
-    case egk = "D2760001448000"
-    case hba = "D27600014601"
-    case smcb = "D27600014606"
-
-    /// 5.3.4 MF / EF.Version2 (SMC-B)
-    /// 5.3.8 MF / EF.Version2 (eGK)
-    /// 5.3.5 MF / EF.Version2 (HBA)
-    /// - Note: for all three the shortFileIdentifier is the same
-    var efVersion2Sfi: ShortFileIdentifier {
-        // swiftlint:disable:next force_unwrapping
-        return EgkFileSystem.EF.version2.sfid!
-    }
-}
 
 extension HealthCardPropertyType {
     static func from(cardAid: CardAid, cardVersion2: CardVersion2) throws -> HealthCardPropertyType {
@@ -49,6 +35,17 @@ extension HealthCardPropertyType {
     }
 }
 
+extension CardAid {
+    /// 5.3.4 MF / EF.Version2 (SMC-B)
+    /// 5.3.8 MF / EF.Version2 (eGK)
+    /// 5.3.5 MF / EF.Version2 (HBA)
+    /// - Note: for all three the shortFileIdentifier is the same
+    var efVersion2Sfi: ShortFileIdentifier {
+        // swiftlint:disable:next force_unwrapping
+        return EgkFileSystem.EF.version2.sfid!
+    }
+}
+
 extension CardChannelType {
 
     var expectedLengthWildcard: Int {
@@ -61,50 +58,42 @@ extension CardChannelType {
     /// Determine `HealthCardPropertyType` either by known initialApplicationIdentifier of the `CardType`
     /// or trying to read EF.Version2.
     /// - Parameters:
-    ///     - writeTimeout: interval in seconds
-    ///     - readTimeout: interval in seconds
-    /// - Returns: Executor that emits a HealthCardPropertyType on successful recognition of the AID and EF.Version2
-    public func readCardType(writeTimeout: TimeInterval = 30.0, readTimeout: TimeInterval = 30.0)
-                    -> Executable<HealthCardPropertyType> {
+    ///     - cardAid: `ApplicationIdentifier` of MF (root) where EF.Version2 is expected to be in.
+    ///                 When now known then the function determines it by itself.
+    ///     - writeTimeout: interval in seconds. Default: 30
+    ///     - readTimeout: interval in seconds. Default: 30
+    /// - Returns: Publisher that emits a HealthCardPropertyType on successful recognition of the AID and EF.Version2
+    public func readCardType(
+            cardAid: CardAid? = nil,
+            writeTimeout: TimeInterval = 30.0,
+            readTimeout: TimeInterval = 30.0
+    ) -> AnyPublisher<HealthCardPropertyType, Error> {
         let channel = self
 
-        return Executable<ApplicationIdentifier>
-                .evaluate {
-                    try channel.card.initialApplicationIdentifier()
-                }
-                .flatMap { (initialApplicationIdentifierData: Data?) in
-                    if let aidData = initialApplicationIdentifierData {
-                        return Executable<ApplicationIdentifier>.evaluate {
-                            try ApplicationIdentifier(aidData)
-                        }
-                    } else {
-                        return try HealthCardCommand.Select.selectRootRequestingFcp(
-                                        expectedLength: channel.expectedLengthWildcard
-                                )
-                                .execute(on: channel, writeTimeout: writeTimeout, readTimeout: readTimeout)
-                                .map {
-                                    let fcp = try FileControlParameter.parse(data: $0.data ?? Data.empty)
-                                    guard let aid = fcp.applicationIdentifier else {
-                                        throw HealthCard.Error.unknownCardType(aid: nil)
-                                    }
-                                    return aid
-                                }
-                    }
-                }
-                .map { (aid: ApplicationIdentifier) in
-                    guard let cardAid = CardAid(rawValue: aid) else {
-                        throw HealthCard.Error.unknownCardType(aid: aid)
-                    }
-                    return cardAid
-                }
+        let cardAidPublisher: AnyPublisher<CardAid, Error>
+        if let cardAid = cardAid {
+            cardAidPublisher = Just(cardAid)
+                    .mapError { $0 as Error }
+                    .eraseToAnyPublisher()
+        } else {
+            cardAidPublisher = channel.determineCardAid()
+        }
+        return cardAidPublisher
                 .flatMap { (cardAid: CardAid) in
-                    try HealthCardCommand.Read.readFileCommand(with: cardAid.efVersion2Sfi,
-                                    ne: channel.expectedLengthWildcard)
-                            .execute(on: channel, writeTimeout: writeTimeout, readTimeout: readTimeout)
-                            .map { response in
+                    HealthCardCommand.Select.selectFile(with: cardAid.rawValue)
+                            .publisher(for: channel, writeTimeout: writeTimeout, readTimeout: readTimeout)
+                            .tryMap { _ in
+                                try HealthCardCommand.Read.readFileCommand(with: cardAid.efVersion2Sfi,
+                                                                           ne: channel.expectedLengthWildcard)
+                            }
+                            .flatMap {
+                                $0.publisher(for: channel, writeTimeout: writeTimeout, readTimeout: readTimeout)
+                            }
+                            .tryMap { response in
                                 let cardVersion2 = try CardVersion2(data: response.data ?? Data.empty)
                                 return try HealthCardPropertyType.from(cardAid: cardAid, cardVersion2: cardVersion2)
                             }
                 }
+                .eraseToAnyPublisher()
     }
 }