+++ ReleaseNotes.md
@@ -1,3 +1,7 @@
+# Release 2.0.2
+- make use of Combine framework
+- add swift package manager support
+
 # Release 1.0.3
 - bugfixes for communication through a secured card channel via NFC - remove embedding of dependencies from project.yml